+++ src/Strategos.Identity.Abstractions.Tests/GlobalUsings.cs
@@ -1,0 +1,16 @@
+// -----------------------------------------------------------------------
+// <copyright file="GlobalUsings.cs" company="Levelup Software">
+// Copyright (c) Levelup Software. All rights reserved.
+// </copyright>
+// -----------------------------------------------------------------------
+
+global using System;
+global using System.Collections.Generic;
+global using System.Linq;
+global using System.Threading.Tasks;
+
+global using NSubstitute;
+
+global using TUnit.Assertions;
+global using TUnit.Assertions.Extensions;
+global using TUnit.Core;